+++ src/components/Controls.js
@@ -1,3 +1,4 @@
+import React from 'react';
 class Controls extends React.Component {
     constructor (props) {
         super(props)
